requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.5
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
spacy==3.7.2
language-tool-python==2.8
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
